# swagger.yaml
openapi: 3.0.3
info:
  title: Energy Dashboard API
  description: |
    Comprehensive API for Energy Dashboard Management System
    
    This API provides endpoints for:
    - Authentication and user management
    - Chart builder functionality
    - Dashboard layout management
    - Data aggregation and analytics
    - Real-time data streaming
    
    ## Authentication
    All endpoints (except login/register) require JWT authentication.
    Include the token in the Authorization header: `Bearer <token>`
    
    ## Response Format
    All responses follow a consistent format:
    ```json
    {
      "success": true,
      "data": {...},
      "message": "Success message",
      "timestamp": "2025-06-18T18:00:00.000Z"
    }
    ```
    
  version: 1.0.0
  contact:
    name: Energy Dashboard API Support
    email: support@energydashboard.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.energydashboard.com/api
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
          description: Indicates if the request was successful
        data:
          type: object
          description: Response data
        message:
          type: string
          description: Response message
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        error:
          type: object
          description: Error details (only present when success is false)
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
...
    ChartConfiguration:
      type: object
      properties:
        id:
          type: string
          description: Chart ID
        name:
          type: string
          description: Chart name
        description:
          type: string
          description: Chart description
        chartType:
          type: string
          enum: [line, bar, column, area, pie, scatter, bubble, spline]
          description: Type of chart
        level:
          type: string
          enum: [SECTOR, STATE, PLANT, EQUIPMENT]
          description: Data aggregation level
        entityId:
          type: string
          description: Entity ID for data filtering
        config:
          type: object
          properties:
            xAxis:
              type: object
              properties:
                field:
                  type: string
                label:
                  type: string
                type:
                  type: string
                  enum: [datetime, number, category]
            yAxis:
              type: object
              properties:
                field:
                  type: string
                label:
                  type: string
                type:
                  type: string
                  enum: [number]
            series:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  field:
                    type: string
                  aggregation:
                    type: string
                    enum: [SUM, AVERAGE, MIN, MAX, COUNT]
                  color:
                    type: string
            filters:
              type: object
            timeRange:
              type: object
              properties:
                start:
                  type: string
                  format: date-time
                end:
                  type: string
                  format: date-time
                preset:
                  type: string
                  enum: [1h, 6h, 24h, 7d, 30d]
        tags:
          type: array
          items:
            type: string
        isTemplate:
          type: boolean
        createdBy:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
            timestamp: "2025-06-18T18:00:00.000Z"
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: "FORBIDDEN"
              message: "Insufficient permissions"
            timestamp: "2025-06-18T18:00:00.000Z"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
            timestamp: "2025-06-18T18:00:00.000Z"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid input data"
              details:
                field: "email"
                message: "Invalid email format"
            timestamp: "2025-06-18T18:00:00.000Z"
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiResponse'
          example:
            success: false
            error:
              code: "INTERNAL_ERROR"
              message: "Internal server error"
            timestamp: "2025-06-18T18:00:00.000Z"
paths:
...
  /chart-builder/charts/{chartId}:
    get:
      tags:
        - Chart Builder
      summary: Get chart by ID
      description: Retrieve a specific chart configuration
      parameters:
        - name: chartId
          in: path
          required: true
          description: Chart ID
          schema:
            type: string
      responses:
        '200':
          description: Chart retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ChartConfiguration'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      tags:
        - Chart Builder
      summary: Update chart
      description: Update an existing chart configuration
      parameters:
        - name: chartId
          in: path
          required: true
          description: Chart ID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                chartType:
                  type: string
                  enum: [line, bar, column, area, pie, scatter, bubble, spline]
                level:
                  type: string
                  enum: [SECTOR, STATE, PLANT, EQUIPMENT]
                entityId:
                  type: string
                config:
                  type: object
                tags:
                  type: array
                  items:
                    type: string
                isTemplate:
                  type: boolean
      responses:
        '200':
          description: Chart updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ChartConfiguration'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      tags:
        - Chart Builder
      summary: Delete chart
      description: Delete a chart configuration
      parameters:
        - name: chartId
          in: path
          required: true
          description: Chart ID
          schema:
            type: string
      responses:
        '200':
          description: Chart deleted successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          message:
                            type: string
                            example: "Chart deleted successfully"
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/ServerError'
...